Address first round of comments
