+++ api/pkg/webhooks/client.go
@@ -64,12 +64,6 @@ const (
 	onErrorAbort  = "abort"
 )
 
-// Config is a helper struct to define the webhook config in a configuration file
-type Config struct {
-	Enabled bool
-	Config  map[EventType][]WebhookConfig `validate:"required_if=Enabled True"`
-}
-
 type WebhookClient interface {
 	Invoke(context.Context, []byte) ([]byte, error)
 	InvokeAsync(context.Context, []byte) error
@@ -82,24 +76,6 @@ type WebhookClient interface {
 
 type simpleWebhookClient struct {
 	WebhookConfig
-}
-
-// WebhookConfig struct is the configuration for each webhook to be called
-type WebhookConfig struct {
-	Name        string `yaml:"name"        validate:"required"`
-	URL         string `yaml:"url"         validate:"required,url"`
-	Method      string `yaml:"method"`
-	AuthEnabled bool   `yaml:"authEnabled"`
-	AuthToken   string `yaml:"authToken"   validate:"required_if=AuthEnabled True"`
-	OnError     string `yaml:"onError"`
-	Async       bool   `yaml:"async"`
-	NumRetries  int    `yaml:"numRetries"`
-	Timeout     *int   `yaml:"timeout"`
-	// UseDataFrom is the name of the webhook whose response will be used as input to this webhook
-	UseDataFrom string `yaml:"useDataFrom"`
-
-	// FinalResponse can be set to use the response from this webhook to the onSuccess callback function
-	FinalResponse bool `yaml:"finalResponse"`
 }
 
 func NoOpErrorHandler(err error) error { return err }
@@ -193,8 +169,8 @@ func validateWebhookConfig(webhookConfig *WebhookConfig) error {
 	if webhookConfig.OnError != onErrorAbort && webhookConfig.OnError != onErrorIgnore {
 		return fmt.Errorf("onError must be either 'abort' or 'ignore'")
 	}
-	if webhookConfig.NumRetries < -1 {
-		return fmt.Errorf("numRetries must be a positive integer or -1")
+	if webhookConfig.NumRetries < 0 {
+		return fmt.Errorf("numRetries must be a non-negative integer")
 	}
 	if webhookConfig.Timeout == nil {
 		def := 10
